Make json object stack explicit.
As of now, the program stack is implicitly used to build the json object
tree. For partial parsing support, we cannot use the program stack as parsing
can be called at anytime. This commit adds a values stack and uses it to
transmit value back along the object tree. All parsing functions now return void
and instead pushes their result on the stack. The functions needing the result
of other parsing functions (mostly for arrays and objects) read values from the
stack and pop them.